+++ builder/clean.sh
@@ -1,6 +1,6 @@
-#!/bin/sh -e
+#!/bin/sh -ex
 #
-# dirclean.sh
+# clean.sh
 # This file is part of Vivarium.
 #
 # Copyright (C) 2019 Jeffery To
@@ -19,8 +19,6 @@
 # along with Vivarium.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-[ "${0%/*}" = . ] || exit 1
-
-. ./clean.sh
-
-rm -rf build_dir/hostpkg staging_dir/hostpkg/*
+rm -fr bin/*
+rm -fr build_dir/*
+rm -fr staging_dir/*